requirements.txt
cloudinary==1.41.0
Django==5.1.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
python-decouple==3.8
python-dotenv==1.0.1